# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Superawesome TGHack project",
    "description": "Superawesome project for hackers",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "api.superawesome.tghack.no",
  "basePath": "/api",
  "tags": [
    {
      "name": "hacker",
      "description": "Endpoints for managing hackers."
    },
    {
      "name": "awesomeHackers",
      "description": "Endpoints for managing CTF categories."
    },
    {
      "name": "flag",
      "description": "Endpoints for managing flag."
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "X-Access-Token",
      "in": "header",
      "description": "This API requires JSON Web Tokens. You acquire your token at login with your user. Unfortunately, this UI does not support API calls. You have to use a tool such as Postman to test the calls."
    }
  },
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "hacker"
        ],
        "description": "Logs into HackBook",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Object of user information for login",
            "schema": {
              "$ref": "#/definitions/requests/LoginRequest"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "$ref": "#/definitions/responses/LoginResponse"
            }
          },
          "401": {
            "description": "Wrong username or password",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          }
        },
        "default": {
          "description": "Unexpected error",
          "schema": {
            "$ref": "#/definitions/responses/ErrorModel"
          }
        }
      }
    },
    "/validateToken/{username}": {
      "post": {
        "tags": [
          "hacker"
        ],
        "description": "Validates a token for a certain username.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "description": "ID of user that gets its token validated",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Validate token successful",
            "schema": {
              "type": "string",
              "example": "OK"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          }
        },
        "default": {
          "description": "Unexpected error",
          "schema": {
            "$ref": "#/definitions/responses/ErrorModel"
          }
        }
      }
    },
    "/AwesomeHackerList": {
      "get": {
        "tags": [
          "awesomeHackers"
        ],
        "description": "Fetches awesome hackers. Requires scope: read:awesomehackers.",
        "responses": {
          "200": {
            "description": "Fetch awesome hacker list successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/responses/AwesomeHackerListResponse"
              }
            }
          },
          "400": {
            "description": "'No awesome hackers found' or 'Unknown error'",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          }
        },
        "default": {
          "description": "unexpected error",
          "schema": {
            "$ref": "#/definitions/responses/ErrorModel"
          }
        }
      }
    },
    "/flag": {
      "get": {
        "tags": [
          "flag"
        ],
        "description": "Fetches the flag. Requires scope: read:flag.",
        "responses": {
          "200": {
            "description": "Fetch flag successful",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "'Unknown error'",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/responses/ErrorMessage"
            }
          }
        },
        "default": {
          "description": "unexpected error",
          "schema": {
            "$ref": "#/definitions/responses/ErrorModel"
          }
        }
      }
    }
  },
  "definitions": {
    "requests": {
      "LoginRequest": {
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "user": {
            "type": "string",
            "example": "maritio_o",
            "uniqueItems": true
          },
          "pwd": {
            "type": "string",
            "example": "Password1234567"
          }
        }
      }
    },
    "responses": {
      "LoginResponse": {
        "type": "object",
        "required": [
          "user",
          "token"
        ],
        "properties": {
          "user": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "expDate": {
            "type": "string"
          }
        }
      },
      "AwesomeHackerListResponse": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ErrorModel": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ResponseMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}